[CI] Parameterize channels in e2e tests by environment variable
We had environment variables before, but nearly all the individual
test invocations hard-coded the "dev" value.

By parameterizing the value everywhere, we unlock the ability to
create ad-hoc runs that can test artifacts from an arbitrary channel,
which in the immediate near term will be useful for helping to
validate core plans refreshes.

We can do more in the future to handle this in a more formal way, but
this is an initial step in that direction.

It should be noted that for general runs of this pipeline, we should
never change the value of the channel; this is *ONLY* for limited
ad-hoc testing.

Signed-off-by: Christopher Maier <cmaier@chef.io>